split sidebar
